Restart service async'y when we're the uninstalled pkg in an upgrade
We have a problem in updating old update-motd packages to a new package.
Some installation scripts run systemctl-try-reload, which runs
update-motd, which runs a system-release hook that asks yum to do
something that requires a lock.  That is bad because a first yum is
causing that sequence and it already owns that lock. Classic deadlock.

We can't just make the new package smarter. Rpm rules are not
well-designed and the last thing to run is the post-un script of the old
evicted package, and if the old package script is determined to do
something bad, the new package can not prevent it.

That is the case here. The already-vended systemrelease and updatemotd
are already going to get in a deadlock. Because we're distributing, in
this package, the software that makes up the middle step in the upcoming
deadlock, we can make it refuse to wait it out.

Closes: SIM#KAROO-1257