Avoid deadlock with systemctlreload-thispackage and Wants: cloud-final-which-called-us-to-reload
Wants: and After: do indeed indicate a soft dependency on the existentiality of
a package (cloud-init), but not on the state of whether it has begun running
and finished running. If it exists at all, then After: will

There is no mechanism built-in for either allowing update-motd not to depend on
cloud-init existing, or asking cloud-init to run things that start update-motd
and not hang on cloud-init having finished.