Don't create an empty /etc/motd.rpmsave
Currently in image recipe builds /etc/motd.rpmsave is *always* present,
which seems a bit superfluous. Change the %post logic to not copy the
file if it's empty.

Check for truthiness from -s or -L as symlinks are treated as non-empty
(their "size" is the number of bytes in the link target).

I also removed the logic of handling the change from using /var/run to
/var/lib/update-motd, as that behavior pre-dates AL2 by a long, long
time.

cr https://code.amazon.com/reviews/CR-3554933